+++ __tests__/reporter/usecase/fetchReporterWithPassword.ts
@@ -1,8 +1,4 @@
-import { ApiKey } from '@libs/apiKey/models/apiKey'
-import {
-  fetchReporterFromApiKeyUseCase,
-  fetchReporterFromPasswordUseCase,
-} from '@libs/reporter/usecase/usecases'
+import { fetchReporterFromPasswordUseCase } from '@libs/reporter/usecase/usecases'
 import { InMemoryReporterClient } from './InMemoryReporterClient'
 
 describe('Reporter usecases', () => {
@@ -16,17 +12,6 @@ describe('Reporter usecases', () => {
 
     expect(fetchedReporter.id).toEqual(1)
     expect(fetchedReporter.nickname).toEqual(username)
-  })
-
-  it('fetch reporter with apiKey', async () => {
-    const client = new InMemoryReporterClient()
-    const fetchReporterWithApiKey = fetchReporterFromApiKeyUseCase(client)
-
-    const apiKey = ApiKey.of('aac4adfd05f0d29832631240764c0d6e15eeac40')
-    const fetchedReporter = await fetchReporterWithApiKey(apiKey)
-
-    expect(fetchedReporter.id).toEqual(1)
-    expect(fetchedReporter.nickname).toEqual('example')
   })
 
   it('empty username', () => {
@@ -63,14 +48,5 @@ describe('Reporter usecases', () => {
     expect(fetchReporterWithPassword(username, password)).resolves.toThrow(
       'Invalid passowrd.'
     )
-  })
-
-  it('invalid api key', () => {
-    const client = new InMemoryReporterClient()
-    const fetchReporterWithApiKey = fetchReporterFromApiKeyUseCase(client)
-
-    const apiKey = ApiKey.of('aac4adfd05f0d29832631240764c0d6e15eeac41')
-
-    expect(fetchReporterWithApiKey(apiKey)).resolves.toThrow('Invalid api key.')
   })
 })
